[fix] autofocus refactor: re-exports, TypeError crashes, MAD=0 outlier bug
- Re-export all functions moved to metrics/focus/integration submodules
  from __init__.py so existing callers are not broken
- Fix run_autofocus returning AutoFocusResult (not float) being used
  as a float in run_multi_position_autofocus, FMAcquisitionWidget, and
  fluorescence_control_widget (_autofocus_worker crash)
- Fix adaptive_frame_integration MAD=0 case silently ignoring
  outlier_threshold by guarding outlier_mask with std_equiv > 0
- Move find_best_focus_plane and calculate_focus_quality to metrics.py
